SubGoalLayout 코드 정리 및 마무리
